Woothash also passes Yves' SMHasher now!

+++ yves/wheathash.h
@@ -19,7 +19,7 @@ static inline uint64_t _wheatmum(const uint64_t A, const uint64_t B) {
 static inline uint64_t _wheatr08(const uint8_t *p){ uint8_t  v; memcpy(&v, p, 1); return v; }
 static inline uint64_t _wheatr16(const uint8_t *p){ uint16_t v; memcpy(&v, p, 2); return v; }
 static inline uint64_t _wheatr32(const uint8_t *p){ uint32_t v; memcpy(&v, p, 4); return v; }
-static inline uint64_t wheathash(const void* key, uint32_t len, uint64_t seed){
+static inline uint64_t wheathash(const void* key, uint64_t len, uint64_t seed){
     const uint8_t *p = (const uint8_t*)key;
     uint32_t i;
     seed += _wheatp1;
